dpyctrl now uses xrandr.
Forcing DPMS off with xset(1) makes all output devices to also goes off. This may be undesiderable for example when you're connected to an HDMI TV and only want to shut your laptop display off. Since xrandr allows to deal on a per-output basis it's easy to solve the problem with it.